+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 2022/04/24
+## [0.3.0] - 2022/04/25
 ### Changed
 - Replace compression code with @s4tk/compression package.
 - Replace cached buffers with CompressedBuffer objects from @s4tk/compression.